Explore: Make Drilldowns box dismissible
